pyproject.toml
[project]
name = "django-boilerplate"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "blacknoise>=1.1.0",
    "dj-database-url>=2.3.0",
    "django-cors-headers>=4.6.0",
    "django-htmx>=1.21.0",
    "django-tailwind>=3.8.0",
    "django-widget-tweaks>=1.5.0",
    "django>=5.1.3",
    "httpx>=0.28.0",
    "pip>=24.3.1",
    "psycopg-binary>=3.2.3",
    "pydantic-settings>=2.6.1",
    "uvicorn>=0.32.1",
]

[dependency-groups]
dev = [
    "daphne>=4.1.2",
    "django-debug-toolbar>=4.4.6",
    "djlint>=1.36.1",
    "ruff>=0.7.3",
    "taskipy>=1.14.0",
]

[tool.taskipy.tasks]
lint = "ruff check . && djlint . --quiet"
format = "ruff check . --fix && ruff format . && djlint . --reformat --quiet"
makemigrations = "python manage.py makemigrations"
migrate = "python manage.py migrate"
run = "python manage.py runserver"
tw_start = "python manage.py tailwind start"
collectstatic = "python manage.py tailwind build && python manage.py collectstatic --noinput -c && python -m blacknoise.compress staticfiles/"


[tool.ruff]
line-length = 79
extend-exclude = ["migrations"]


[tool.ruff.lint]
preview = true
select = ["A", "B", "C", "E", "F", "I", "PL", "PT", "Q", "S", "T", "W", "DJ"]
ignore = ["E501", "PLR6301"]

[tool.ruff.lint.flake8-quotes]
inline-quotes = "single"

[tool.ruff.format]
preview = true
quote-style = "single"


[tool.djlint]
max_blank_lines = 1
indent = 4
tabs = true
format_css = true
format_js = true
use_gitignore = true
space_around_blocks = true
blank_line_before_tag = "include,block,endblock,crispy,if,else,endif,for,endfor,comment,endcomment,with"
blank_line_after_tag = "include,block,endblock,crispy,if,else,endif,for,endfor,comment,endcomment,with,load,extends"
ignore = "H005,H006,H007,H008,H009,H010,H011,H012,H013,H014,H015,H016,H017,H018,H019,H020,H021,H022,H023,H024,H025,H026,H027,H028,H029,H030,H031,H033,H035,H036,H037,J004,J018,T001,T002,T003,T027,T028,T032,T034"

[tool.djlint.css]
indent_size = 4

[tool.djlint.js]
indent_size = 4
